+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.0] - unreleased
 
+### Fixed
+
+- Handling of space character in path names for included files.
+
 ## [1.1.0] - 2024-12-21
 
 ### Added